Per reviewer comments, get multi-chunk memory from the sublocales.
I hadn't coded this up yet because we haven't decided what we'll
actually need at this low a level in the stack.  Obviously
  on <sublock> { ... allocate ... }
is the right thing in user space, but we're much lower than that here
and might want to be more specific.  Nevertheless it's good to say we
want sublocale-resident memory for each chunk, and putting it in the
code is better than putting it in, say, a comment.