Handle generic types with default in function return
Gets test/functions/ferguson/return-type-generic-default.chpl and
test/parallel/sync/bradc/testBadSyncSingle working again by explicitly
handling the case that a function returns a generic type. Before this PR,
type aliases were always concrete. After this PR, they can continue to be
generic. That means they could be returned from a function. But a
function must return a concrete type.

In particular, adjusts resolveDefaultGenericTypeSymExpr to return a Type*
and to handle the case of figuring out the type of a FLAG_TYPE_VARIABLE
VarSymbol (which is how we represent a type alias). It returns a Type*
for convenience.

resolveDefaultGenericTypeSymExpr is called (as it was before) in both
resolveSpecifiedReturnType and resolveCoerce. Returning the type in
resolveDefaultGenericTypeSymExpr simplifies its use in
resolveSpecifiedReturnType.