Changed from busy-wait to waitFor()
In conversations with the Bens, particularly BHarsh, I took an IOU to
see how much switching from a busy-wait to a more principled waitFor()
hurts performance.  It's reasonably minimal, so I made the switch.  In
addition to using supplied features and avoiding the taboo of
busy-waiting, this also reduces the need for the helper functions which
I eliminated.