Permit 'param' decls sans initializers
While it's not very interesting or useful, it seems that we should
permit 'param' declarations without initialization expressions for
orthogonality (where the 'param's value would be its default value).
It turns out that it was a five-minute fix to enable this and that
only one test seemed to rely on the old "error message" behavior,
so I made the fix and propose we accept it.