Update spec w.r.t. const/param and initializers
I went looking in the spec to see if it implied that params require
initializers, but it didn't seem to.  However, constants claimed that
they needed initializers even though the implementation doesn't
require it (and shouldn't, in my opinion).  Since this is thematically
related to this PR, I included it here.