Flag 'param's of illegal type
In the past, we haven't indicated whether a 'param' declaration is of
legal type or not -- instead, it's generated an error message due to
the initializing expression not being a 'param'.  This seems a bit
roundabout, so this commit attempts to improve things by first
inspecting the type to see whether it's one that we consider to be a
supported param type (currently, bools, ints, uints, reals, imags,
enums, and strings).

I added a test to lock in this behavior and, in doing so, noticed/
remembered that we don't support 32-bit params because we don't have a
defaultOf() routine that can initialize them to a param value.  Filed
a separate future against this called param32BitFloat.chpl, but kept
my compiler test for param types calling 32-bit floats legal because
that's what I think we want (and because everything downstream of the
default init can handle it).