A variety of fixes after "stress" testing.
I tested with param -s defRectArrMultiDDataSizeThreshold=32 to stress
the code.  This forced many more of the tests to use arrays stored in
multi ddata chunks than in the default (2**20) case.  The result was
about 100 errors.  Here several of those are fixed.

==========
modules/internal/DefaultRectangular.chpl

Instead of putting vectors of chunked ddata and shifted-ddata in the
class, put a vector of per-chunk information that includes not only the
data and shiftedData chunks but also the chunk's subrange of the dim
over which we're chunked, and the data offset of this chunk's data.
Make corresponding changes throughout the modules, but especially in
dsi{Reindex,Slice,Rankchange,Reallocate}.  Fix a number of bugs in those
latter functions.  Simplify the addressing logic and correct some bugs
there as well.

==========
modules/dists/BlockDist.chpl
modules/dists/CyclicDist.chpl
modules/dists/StencilDist.chpl

Make changes corresponding to those in DefaultRectangular, plus also
remove some no-longer-needed code that I'd missed before.