Fix array I/O for multi-ddata.
Whole-array I/O with multi-ddata didn't work at all.  Apparently this
was an area where I did some stopgap work and left that as a marker, but
never came back and finished it.  Fix it up and get it working.

While here (in DefaultRectangularArr), clean up some type problems wih
unsigned int index types in a couple of spots.

In the tests, fix a couple more that need .dataChunk(0) in place of
.data, and assert()s to ensure they're only run with single-ddata
arrays.