Remove problematic "local field" pragmas.
The pragma "local field" decorations on the data and shiftedData fields
in the _multiData record are causing errors which didn't occur when the
corresponding datavec and shiftedDataVec members were directly in the
DefaultRectangularArr class.  I'm not sure what the problem is, but for
now just comment them out, mentioning why.