Fix sloppiness in bulk transfer.
The transfer code was making assumptions that weren't warranted, given
what doiUseBulkTransfer() said could be handled.  In the particular case
of optimizations/bulkcomm/alberto/Block/3dStrideTest, the affected chunk
was different for the A and B arrays, but doiBulkTransfer() wasn't coded
to handle that.  Fix doiUseBulkTransfer() to be more precise about what
we can deal with, and doiBulkTransfer() to actually handle everything we
can.