Add a few record initializers tests, to verify copy initializer behavior
checking-copy-init.chpl provides two init methods: one which takes no arguments
and one which takes only another instance of the type (aka, a copy initializer).
This test passes.  checking-copy-init-bad.chpl only provides the first init
method (the no arguments version).  It fails when resolving the useIt()
function, as that function returns an instance of that type, causing a copy to
occur.  record-with-no-arg-init.chpl only provides the first init method as
well, but does not have a useIt() function.  It fails when trying to resolve
the initCopy and autoCopy for that type, during resolution's "resolveAutoCopies"
function, even though I believe no copy is actually occuring.