More test modifications based on discussion with Michael
Add tests for actual copies being made by the user, including one where the
copy initializer was a Phase 2 body and one where the copy initializer was a
Phase 1 body.  No longer expect the move case to generate an error message when
a copy initializer was not provided (it should just work) - this might be turned
into a future if the change is too tricky.