Two changes to function resolution and auto copies
1) Move the autoCopyMap internal to function resolution, and replace all but
one or two gets from it with an accessor function call.  Also replace the
get_keys call in checks and a test-and-use in lowerIterators with similar
accessor calls.

2) modify resolveAutoCopyEtc to behave differently when the type has defined an
initializer.  In this case, it is possible that the initCopy and autoCopy
functions will not appropriately resolve, as they expect to access an
initializer that is callable with all fields as arguments and that is not
created by default when the user defines their own initializer.  So, try to
resolve the expected call to the copy initializer and if that doesn't work,
don't try to resolve the autoCopy for that type.

Saving my progress for tonight.