chplspell: replace use of shopt -u globstar
bash prior to version 4.0 doesn't support the globstar option.  Macs
only have bash 3.2 installed by default.  So replace the use of
globstar globbing with a $(find) command.

The file lists produced by find are sorted by directory, so we don't
need to sort them any more.

The sorting was the only place the globs in DEFAULTFILES[] were
resolved, so those need to be explicitly resolved now, after cd'ing to
$CHPL_HOME.