Fix attempt to access autoCopy when no copies are made
This was causing test/classes/initializers/records/record-with-no-arg-init.chpl
to encounter issues - prior to hiding the autoCopyMap, we would get a segfault
because I had inserted NULL as a sentinel value for types whose autoCopies would
not resolve.  After the autoCopyMap was hidden and the .get calls replaced with
an accessor that would complain if you tried to retrieve a value that was NULL,
we would encounter problems because of this check function.  Now I eliminate
the access that was going to return NULL by checking if the key has a
corresponding value first before trying to obtain it.