Add some .futures about type methods
Type methods currently accept any type as the receiver, so defining a type
method with the same name in two different classes is ambiguous instead of
using the type of the receiver to disambiguate.

Also, a type method defined in class "C" can be called on any type, for example
"int".

Both of these can be worked around with simple where clauses.