Update isDefAndOrUse to handle PRIM_SET_MEMBER/PRIM_SET_SVEC_MEMBER
A long-standing issue with this code was that it did not correctly
handle primitives setting a field in a record.

In order for reaching-definitions analysis to make sense,
we need PRIM_SET_MEMBER to count as both a def and a use.

Consider a record with two fields, x and y.

If we have the code

  def TMP
  move TMP, defaultValue
  set-member TMP, x, 1
  ... use TMP ...

the set-member modifies TMP but does not totally replace
the previous definiton. If we considered it to be just a "def",
reaching definitions + dead code elimination could remove the
move TMP, defaultValue since the set-member should "define" TMP.