Fix problems with ref-pair and _this intent
Before this change, refNecessary in cullOverReferences included a special
case to always consider the '_this' argument to be passed by ref. This
was added in commit f2bc2b27 which dates back to when 'setter' was
initially added.

However, methods on an array or domain "wrapper record" should not
be considered to set the array or domain wrapper record. (They *can*
set array elements or re-assign a domain. But since these wrapper records
are immutable, it probably does not make sense to have them pass
by ref).

To resolve the issue, make 3 related changes:

 1) Remove the special case code in refNecessary. The intent for the
    _this argument computed in resolveIntents will be used instead
    of always considering _this to be passed by ref.
 2) In resolveIntents, pass the wrapper record for domain/array/dist
    as 'const ref' since these wrapper records are immutable.
 3) In buildDefaultFunctions, mark "setter" accessor functions for
    records as taking in _this by REF.

Taken together, these changes resolve the future
 test/functions/bharshbarg/arr-ref-return.chpl