Add 5 new initializers tests, which occurred to Mike and I yesterday
explicitAssignment.chpl validates the behavior when a generic type instantiation
is used as the declared type for a variable we are creating, but the initializer
it calls would otherwise have created an instance of a different instantiation.
Mike and I think that if the type for the second instantiation can be
coerced/cast into the type for the first, then it should be okay to assign from
an instance of the second into the first.  However, we don't expect this to
work right now (and probably not for a while).

The remaining tests validate the behavior with respect to secondary
initializers.  bothSecondaryAndPrimary.chpl defines both a secondary and a
primary initializer for the type (and the secondary initializer is also defined
in the same module as the type itself), and ensures that both can be found.  The
other three include no primary initializer.  inDiffModule.chpl validates the
behavior when the secondary initializer is defined in a different module than
the one where the type is defined and an instance of the type is created in the
module where the original type is defined only (i.e. the module with the
secondary initializer is not used).  inDiffModule2.chpl validates when the
secondary initializer is defined in a different module than the one where the
type is defined and an instance of the type is created in the same module as
the secondary initializer only.  I do not have a test when an instance of the
type is created in both the module where the type is defined and in the module
where the secondary initializer is defined, as that runs into some interesting
point of instantiation semantics that I didn't want to dive into just yet.
inSameModule.chpl validates the behavior when the secondary initializer is
defined in the same module as the type on which it is defined.