Move the multi-ddata metadata fields, checking cache effects for 'flat'.
I'm temporarily moving the new multi-ddata metadata fields in the
DefaultRectangularArr class so that they follow the pre-existing data
and shifted_data fields instead of being before them.  I'm doing this to
see if the small performance degradation we saw with locModel=flat for
some tests is due to the new fields interfering with cache line sharing
effects.  (Manual runs for several of the tests that seem to show solid
~5-8% degradation since the multi-ddata PR are not reproducing that
degradation, thus the experiment.)