Remove the old flag FLAG_OMIT_FROM_CONSTRUCTOR
I ran across this flag when trying to understand type constructors.  It appears
to be checked against but not attached to anything in either the compiler or
the modules.  It also appears to have been around for a really long time.  I
suspect that the last set of it was removed some time after 2010 (the latest
time the flags were reshuffled), but I don't know when that was removed.  In
any case, we don't use it now, so we might as well remove it.