Adding API for Task ID comparison and conversion to string
API is required for comparing task IDs that are not integers,
since other types such as struct wont work with binary operators.
Also adding a 'to string' API to help with printing of task IDs.
Adding CHPL_TASK_ID_STRING_MAX_LEN to define the
maximum length of string buffer to use.