Handle types that cannot be copied in new initializers story
Types that cannot be copied are important for implementing
an equivalent to C++'s std::unique_ptr.

This patch works around the issue that the compiler adds and resolves
calls to initCopy before these calls might be eliminated.  Thayt way,
types that cannot be copied should still work in case the initCopy calls
are removed after resolution. In particular, callDestructors removes
extra initCopy calls for certain common patterns such as capturing
the returned value from a function into a variable.

In order to do that:
 * adjusted generic instantiation to call fixupDefaultInitCopy when
   instantiating a compiler-generated initCopy function
 * implemented fixupDefaultInitCopy to explicitly handle the creation
   of the initCopy function for types with initializers defined.
   In particular, if a copy-initializer is not available, create a "dummy"
   initCopy marked with the flag FLAG_ERRONEOUS_INITCOPY. Otherwise,
   create an initCopy that runs the copy-initializer.
 * at the end of callDestructors, add code to check that no calls
   to initCopy functions marked with FLAG_ERRONEOUS_INITCOPY remain.
   If any do, raise a user-facing compilation error.

In order to get all of that to work, I also needed to adjust
buildDefaultFunctions to avoid generating a compiler-default initCopy for
a record that defines an initializer.