Add test showing cases that should insert cpp lines
This test tries to enforce that the compiler inserts #line directives
when --cpp-lines is thrown (whether by a user or developer, whether C
is being saved or not) and when -g is thrown by users (but not by
developers).  It seems that previous attempts to clamp down on when we
beautify (which is also where we insert #line directives currently)
resulted in not respecting the --cpp-lines flag, which I think it a
mistake.  It may have been postulated that the user wouldn't know
whether or not we added #line directives if they did not save the C
code, but this is not the case.

For example, if the back-end C compiler hits a compilation error when
compiling with --cpp-lines even if the C is not saved, it should
reveal the .chpl line number where the problem occurred rather than
the .c line number.  This can be helpful for quickly tracking down the
root cause of the problem without looking through all the C code.
Another use case is when the __FILE__ or __LINE__ directives are used
in the generated C code, as in this test.

One downside of this test is that it does not check that other cases
(e.g., not throwing --cpp-lines, compiling -g as a developer) do not
generate cpp-lines.  That's harder to do since the C files can be
saved in different places and therefore have different __FILE__ names
and also because the __LINE__ printed would be unstable as the
generated code changes over time.

As noted in util/files.cpp, beautify and the insertion of #lines are
arguably interwoven more tightly than they should be.  E.g., if
--savec is not used, we could insert #lines without beautifying
(indenting/ pretty printing) the generated code.  I didn't undertake
any improvements here as part of this effort.