Skip new mandelbrot version for linux32
This change follows the lead of PR #4405 which disabled some versions
of mandelbrot for linux32 and --fast.

Over the weekend, testing of my new mandelbrot version failed on
linux32 due to the absence of a single bit on the uppermost border
of the mandelbrot set, which can be seen by zooming into the
expected solution on the CLBG site:

  http://benchmarksgame.alioth.debian.org/download/mandelbrot200.png

It's a little odd that this bit is "on" given how far it is outside
of the mandelbrot set, and on linux32, it's computed as being "off".
But only when --denormalize is on (as it is by default, currently).

As noted on the associated JIRA issue
(https://chapel.atlassian.net/browse/CHAPEL-244), some of the
reference versions of mandelbrot also seem to fail on the CLBG site's
32-bit runs, so perhaps this is not unexpected, but it's still
disconcerting.

We didn't seem to get failures for --fast on chapcs (as reported on
the links above) nor for cygwin32, so I made my .skipif more
specialized than those for now.  If the behavior is unstable over
time, it can be expanded as necessary.