Convert array init method from config var to config param
config vars show up in the executables --help output, which isn't what we want
since this isn't intended to be user facing. So instead just make
arrayInitMethod a non-config global var, and make a config param
defaultArrayInitMethod. This makes things a little more complicated, but again
it's not user facing, so no big deal.