Disable CHPL_UNWIND during start_test runs
This commit disables CHPL_UNWIND during start_test runs, since the
output can be dramatically different (e.g., if an expected error
message is called, which calls halt, a stack trace is printed.

It's an open question as to how we'd do testing in a "CHPL_UNWIND on
by default" world, but in the meantime, I'm trying to work with
CHPL_UNWIND on by default to see how well it works and when it seems
to get confused (or get me confused).  In the meantime, this seems
like the nice thing to do for developers who use it.