Remove --fast skipif for 32-bit mandelbrot
I was thinking of this skipif as skipping the test for --fast testing,
but of course it also skips it for performance testing, which is what
I want to be doing.  Since the --fast failure didn't seem to occur
for any of the nightly testing configurations that we run, it seems
OK to disable it.  (It would presumably fail if we did --fast
performance testing on chapcs or did stricter verification of the
performance-sized mandelbrot output... unless the glitch doesn't
occur at that size?)