Fixed minor bug in printResolutionErrorUnresolved
It used to be the case that printResolutionErrorUnresolved()
reached this code at the end of it:

  if( developer ) {
    // Should this be controlled another way?
    USR_FATAL_CONT(call, "unresolved call had id %i", call->id);
  }
  USR_STOP();

without actually issuing a USR_FATAL_CONT(), for example when
calling these() method with arguments.

If so, no "unresolved" error would be issued and
* with --devel, only ""unresolved call had id NNN" would be printed
* with --no-devel, printResolutionErrorUnresolved() would return
and resolveNormalCall() would hit INT_FATAL(call, "unable to resolve call").

This PR fixes that by tracking the cases not yet handled.
It would be eaiser to just check exit_eventually instead,
however unfortunately exit_eventually is static to misc.cpp.

While there, I replaced USR_FATAL_CONT(call, "unresolved call had id %i",...)
with USR_PRINT because this is an information statement
and having it cause an error is not/should not be necessary.