Create same number of mem descs w/ and w/o --devel
MOTIVATION

I found the number of AST nodes created with and without --devel
to differ. Here is what I believe is the cause.

callChplHereAlloc() invokes newMemDesc() passing it the name of
the type being allocated. newMemDesc() creates a new_IntSymbol()
(new VarSymbol + new DefExpr) for each new name that it gets.

When compiling with --no-devel, type names for domains/arrays are
made nicer, e.g.:

  [CSRDom(2,int(64),domain(2,int(64),false))] int(64)

Cf. with --devel they are more verbose, e.g.:

  BaseSparseArrImpl(int(64),2,int(64),CSRDom(2,int(64),_domain(DefaultRectangularDom(2,int(64),false))))

In a couple of cases, such as this, two different types have the same
(nicer) name with --no-devel whereas different names with --devel. So
the latter causes extra new_IntSymbol() and so extra two nodes
compared to the former.

I only looked for differences through the first half of resolution.
I did not compare the number of nodes created downstream from there.

To observe the difference, I ran the following in the debugger:

   chpl $CHPL_HOME/test/users/engin/sparse_bulk_dist/sparseBulk.chpl -ssparseLayoutType=CSR

Then, break on printCallStackOnError() and print uid.

Before this change, uid is smaller by 6 when running with --no-devel
than with --devel. After this change, it's the same.

IMPLEMENTATION

As suggested by Michael, I made newMemDesc() create separate
descriptor numbers based on the type, rather than its name.

* This required a deviation from name-centric memory descriptors.
In a few cases newMemDesc() is invoked on char* literals e.g.
"global heap-converted data", "bundled args", etc.
Prior to this change, there was a 1:1 between the name and
the descriptor number. Now the same type name can have several
descriptors if multiple Types have the same name. See the above
example.

* To implement that, I created a separate newMemDesc() and memDescsMap
that operate on Symbols. Note that I did caching on Symbols' ids,
not pointers. This is to guard against the case where a Symbol
is removed from tree and its pointer is reused. Even though such
a case is unlikely.