Allow domain implementations to write domain assignment.
Before this change, domain assignment was implemented in ChapelArray and
included basically two branches: one branch for regular domains that
amounted to calling setIndices(getIndices) and a second branch for
irregular domains that worked in serial and used
dsiIndsIterSafeForRemoving.

This implementation approach presents problems for things like
distributed associative or distributed sparse, where it's important that
those implementations have an opportunity to optimize. In particular,
assigning one distributed associative domain to another, where they share
the same distribution, is an important case to optimize for. In that
case, the domain assignment should be possible to implement with no
communication (assuming privitaization and no arrays over the
lhs-domain). Since such implementations depend on the details of the
implementation, it makes sense to have each domain map define
how domain assignment is to be accomplished.

This strategy is complicated by the fact that domain assigment is
sometimes possible between domains of different types. A typical exampe
might be assignment between a distributed domain and a non-distributed
one. Of course, one could argue that such things such things should be
compilation errors... but the pattern itself is useful and currently
supported directly in domain assignment.

So, the new method is something like this:

  proc SomeDom.dsiAssignDomain(rhs: domain, lhsPrivate:bool) { }

The first argument is generic over all domains.
The second argument is used to optimize initCopy-like cases
when assigning domains.

Now, note that the DSI interface is partially specified in the methods in
the base class BaseDom. It would be nice to put a dsiAssignDomain there,
but doing so encounters a problem with how generic functions interact
with virtual methods.

When a generic function is overridden (such as with one taking a `domain`
argument like dsiAssignDomain), the compiler resolves versions of that
functions for each instantiation up and down the inheritance tree.

That means that if we have

  proc BaseDom.dsiAssignDomain(rhs:domain)

and somewhere there is

  proc DomainOne.dsiAssignDomain(rhs:domain)
  proc DomainTwo.dsiAssignDomain(rhs:domain)

and then there is a call like

  someDomainOne.dsiAssignDomain( some2DDomain )

then function resolution will instantiate BaseDom.dsiAssignDomain
for a that particular 2D domain argument. Then, it will instantiate
DomainTwo for that particular 2D domain argument. But DomainTwo
isn't necessarily one that can be set to a 2D domain - it might be 1D.

I think what the compiler is doing here is reasonable because if
BaseDom includes dsiAssignDomain, one could presumably call it
with a 2D domain type and expect virtual dispatch to handle it.

To do better at describing and enforncing the DSI "interface" at compile
time, I think we need interfaces/concepts. For now, I'll seek to document
this required method in the DSI technote.