Respond to reviewer feedback
At Vass's suggestion, add INT_FATAL if checkonly is true in
resolveUninsertedCall's else branch.  I believe there is no reason we couldn't
support checkonly for the other cases, but it seemed unnecessarily burdensome to
thread the boolean through all the other functions called by resolveCall in the
primitive case.  This error will alert future developers to the boolean being
dropped on the floor, so that they can decide for themselves whether it is
worthwhile to support in those cases.