Fix bug with .c_str()
The variable struct_name vwas getting corrupted in some
cases. It was the result of .c_str() on a string which isn't
necessarily well defined if arbitrary code runs before the
value is used, which was the case.

Resolved the issue by using astr() to add the string to the Chapel
table.