Using API to compare and print Task ID
Follow up on Pull requests #5152 and #5177.
The task IDs are used like integers, for example
compare using binary operators and print using %d
format specifier. This will not work when task ID
is of non integer type such as struct. Therefore
use API to do comparison and conversion to string.