Minor cleanup.
While there, I modified callChplHereAlloc() to always accept Type*
instead of Symbol*, thus avoiding Type*->Symbol*->Type* conversion.

The only place where this was non-trivial was replacement:
  callChplHereAlloc(fn->_this) -> callChplHereAlloc(fn->_this->type)
in
  build_constructor(AggregateType* ct)

BTW in my observation this allocates 'this' of the type '_ref',
which is generic. I did not investigate what becomes of this allocation
when the constructor for _ref is instantiated, if it ever is.
This seems to be the first call to newMemDesc, so the memory descriptor is 0.

Other "while there":

* Removed a couple of unnecessary lines from callChplHereAlloc().

* Removed an unnecessary SET_LINENO() call from scopeResolve().

* I also considered removing the optional 'md' formal in callChplHereAlloc()
because it is currently unused. I chose to leave it there just in case
and for symmetry with insertChplHereAlloc.