Intra-locale parallelism for BlockCyclic
Prior to this change, BlockCyclic would only create one task per locale
when iterating over the domain (or array) in parallel. Iterating in
parallel over ``myStarts`` yields positive results for ptrans on
16-node-xc, and seems like a good starting point.

In support of this change, the ``tasksPerLocale`` argument to the
Blockcyclic constructor has been renamed to ``dataParTasksPerLocale`` to
match other distributions.

The parallel iterator bodies for BlockCyclicDom and BlockCyclicArr were
also cleaned up by:
- removing old commented-out debugging writelns
- caching common expressions to reduce verbosity