Detect GSL in test
For now, this just follows the pattern from the ZMQ SKIPIF which is to
use Python's ctypes find_library. That finds the .so file which isn't
quite what we need but is a good start.

For now, I think it's a reasonable target that all of our nightly
testing systems either:
 - have ZMQ/GSL .so .a and .h files available, OR
 - do not have ZMQ/GSL .so installed

Doing a better job at this is something we absolutely ought to do
but it's not clear to me how to do it easily and correctly.

This should be merged after PR #5205.