Param guard reads/writes of chpl_arrayInitMethod
chpl_arrayInitMethod is a global variable. Reading it required a get, which had
the nasty side effect of increased comm counts and not being able to declare
arrays inside of a local-block.

This adds a param guard around it, which avoids the gets, but makes changing
the array init method at runtime more cumbersome.

Now instead of just doing

    chpl_arrayInitMethod = ArrayInit.parallelInit;
    var A: [1..100] int;

you now have to compile with -schpl_arrayInitMethodRuntimeSelectable and do

    chpl_setArrayInitMethod(ArrayInit.parallelInit);
    var A: [1..100] int;

which is annoying, but fine since this feature is just meant to help us explore
initialization strategies effect on first-touch.